// src/GeneratedClient/models/MessageType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WaitingList
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const MessageType = {
    Success: 'success',
    Info: 'info',
    Warning: 'warning',
    Error: 'error'
} as const;
export type MessageType = typeof MessageType[keyof typeof MessageType];


export function instanceOfMessageType(value: any): boolean {
    for (const key in MessageType) {
        if (Object.prototype.hasOwnProperty.call(MessageType, key)) {
            if (MessageType[key as keyof typeof MessageType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function MessageTypeFromJSON(json: any): MessageType {
    return MessageTypeFromJSONTyped(json, false);
}

export function MessageTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessageType {
    return json as MessageType;
}

export function MessageTypeToJSON(value?: MessageType | null): any {
    return value as any;
}

export function MessageTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): MessageType {
    return value as MessageType;
}