// src/GeneratedClient/models/PartyModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WaitingList
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WaitingListModel } from './WaitingListModel';
import {
    WaitingListModelFromJSON,
    WaitingListModelFromJSONTyped,
    WaitingListModelToJSON,
    WaitingListModelToJSONTyped,
} from './WaitingListModel';

/**
 * 
 * @export
 * @interface PartyModel
 */
export interface PartyModel {
    /**
     * 
     * @type {string}
     * @memberof PartyModel
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof PartyModel
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof PartyModel
     */
    size?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PartyModel
     */
    isReady?: boolean;
    /**
     * 
     * @type {WaitingListModel}
     * @memberof PartyModel
     */
    waitingListModel?: WaitingListModel;
}

/**
 * Check if a given object implements the PartyModel interface.
 */
export function instanceOfPartyModel(value: object): value is PartyModel {
    return true;
}

export function PartyModelFromJSON(json: any): PartyModel {
    return PartyModelFromJSONTyped(json, false);
}

export function PartyModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): PartyModel {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'size': json['size'] == null ? undefined : json['size'],
        'isReady': json['isReady'] == null ? undefined : json['isReady'],
        'waitingListModel': json['waitingListModel'] == null ? undefined : WaitingListModelFromJSON(json['waitingListModel']),
    };
}

export function PartyModelToJSON(json: any): PartyModel {
    return PartyModelToJSONTyped(json, false);
}

export function PartyModelToJSONTyped(value?: PartyModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'size': value['size'],
        'isReady': value['isReady'],
        'waitingListModel': WaitingListModelToJSON(value['waitingListModel']),
    };
}